\name{globtemp}
\alias{globtemp}
\docType{data}
\title{
       The First Differences of Annual Golbal Surface Air Temperatures from 1880-1985
}
\description{
The first differences of annual global surface air temperatures in Celsius from 1880 through 1985.
}
\usage{data(globtemp)}
\format{
  A matrix containing 106 observations. The first column is the time from 1880-1985. The second column is the first difference of annual global surface air temperature in celsius. 
}
\details{
The first differences of the annual global surface air temperatures in Celsius from 1880 through 1985.
}
\source{
\url{http://datamarket.com/data/set/22ku/annual-changes-in-global-temperature-1880-1985#!ds=22ku&display=line}
}
\seealso{
 \code{\link{qn}}  \code{\link{ma.filter}}
}
\examples{
data(globtemp)
globtemp
}
\keyword{datasets}
